Exclude training on image build
